@@ -2,23 +2,23 @@ name: flutter_landingpage
 description: app 介绍性网页.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.0.0+1
+version: 1.2.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: '>=2.19.2 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.5
   url_launcher: ^6.0.10
-  dio: ^4.0.0
+  dio: ^5.3.0
   shared_preferences: ^2.1.1
   
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^2.0.2
 
 flutter:
   uses-material-design: true
