@@ -1,4 +1,4 @@
 class Api{
   static String host="https://xx.com";
-  static String getHome = host+"/gethome";
+  static String getHome = "$host/gethome";
 }
