@@ -7,12 +7,10 @@ class OneHeaderR extends StatelessWidget {
   String assetPath;
   final bool fixedSize;
 
-  OneHeaderR({required this.assetPath, this.fixedSize = true});
+  OneHeaderR({super.key, required this.assetPath, this.fixedSize = true});
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      child: Image.asset("assets/" + assetPath),
-    );
+    return Image.asset("assets/$assetPath");
   }
 }
